sitemotorcycles/requirements.txt
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
decorator==5.1.1
defusedxml==0.8.0rc2
Django==5.0.3
django-environ==0.11.2
django-extensions==3.2.3
exceptiongroup==1.2.0
executing==2.0.1
gmane==0.1.dev25
hiredis==2.3.2
idna==3.6
ipython==8.22.2
jedi==0.19.1
kombu==5.3.6
matplotlib-inline==0.1.6
networkx==3.2.1
oauthlib==3.2.2
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social==0.1.dev3
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
splinter==0.21.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
django-celery-results==2.5.1
djangorestframework==3.15.1
Flask==2.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
Werkzeug==3.0.1
